--- a/src/UglyToad.PdfPig.Core/IInputBytes.cs
+++ b/src/UglyToad.PdfPig.Core/IInputBytes.cs
@@ -47,8 +47,7 @@
         /// Fill the buffer with bytes starting from the current position.
         /// </summary>
         /// <param name="buffer">A buffer with a length corresponding to the number of bytes to read.</param>
-        /// <param name="length">Optional override for the number of bytes to read.</param>
         /// <returns>The number of bytes successfully read.</returns>
-        int Read(byte[] buffer, int? length = null);
+        int Read(Span<byte> buffer);
     }
 }