--- a/src/UglyToad.PdfPig.Core/Polyfills/StreamExtensions.cs
+++ b/src/UglyToad.PdfPig.Core/Polyfills/StreamExtensions.cs
@@ -6,19 +6,37 @@ using System.Buffers;
 
 internal static class StreamExtensions
 {
-    public static void Write(this Stream stream, ReadOnlySpan<byte> data)
+    public static void Write(this Stream stream, ReadOnlySpan<byte> buffer)
     {
-        var buffer = ArrayPool<byte>.Shared.Rent(data.Length);
+        var tempBuffer = ArrayPool<byte>.Shared.Rent(buffer.Length);
 
-        data.CopyTo(buffer);
+        buffer.CopyTo(tempBuffer);
 
         try
         {
-            stream.Write(buffer, 0, data.Length);
+            stream.Write(tempBuffer, 0, buffer.Length);
         }
         finally
         {
-            ArrayPool<byte>.Shared.Return(buffer);
+            ArrayPool<byte>.Shared.Return(tempBuffer);
+        }
+    }
+
+    public static int Read(this Stream stream, Span<byte> buffer)
+    {
+        var tempBuffer = ArrayPool<byte>.Shared.Rent(buffer.Length);
+
+        try
+        {
+            int read = stream.Read(tempBuffer, 0, buffer.Length);
+
+            tempBuffer.AsSpan(0, read).CopyTo(buffer);
+
+            return read;
+        }
+        finally
+        {
+            ArrayPool<byte>.Shared.Return(tempBuffer);
         }
     }
 }