--- a/src/UglyToad.PdfPig.Fonts/CompactFontFormat/CompactFontFormatData.cs
+++ b/src/UglyToad.PdfPig.Fonts/CompactFontFormat/CompactFontFormatData.cs
@@ -11,7 +11,7 @@
     /// </summary>
     public class CompactFontFormatData
     {
-        private readonly IReadOnlyList<byte> dataBytes;
+        private readonly ReadOnlyMemory<byte> dataBytes;
 
         /// <summary>
         /// The current position in the data.
@@ -21,13 +21,13 @@
         /// <summary>
         /// The length of the data.
         /// </summary>
-        public int Length => dataBytes.Count;
+        public int Length => dataBytes.Length;
 
         /// <summary>
         /// Create a new <see cref="CompactFontFormatData"/>.
         /// </summary>
         [DebuggerStepThrough]
-        public CompactFontFormatData(IReadOnlyList<byte> dataBytes)
+        public CompactFontFormatData(ReadOnlyMemory<byte> dataBytes)
         {
             this.dataBytes = dataBytes;
         }
@@ -93,12 +93,12 @@
         {
             Position++;
 
-            if (Position >= dataBytes.Count)
+            if (Position >= dataBytes.Length)
             {
-                throw new IndexOutOfRangeException($"Cannot read byte at position {Position} of an array which is {dataBytes.Count} bytes long.");
+                throw new IndexOutOfRangeException($"Cannot read byte at position {Position} of an array which is {dataBytes.Length} bytes long.");
             }
 
-            return dataBytes[Position];
+            return dataBytes.Span[Position];
         }
 
         /// <summary>
@@ -106,7 +106,7 @@
         /// </summary>
         public byte Peek()
         {
-            return dataBytes[Position + 1];
+            return dataBytes.Span[Position + 1];
         }
 
         /// <summary>
@@ -114,7 +114,7 @@
         /// </summary>
         public bool CanRead()
         {
-            return Position < dataBytes.Count - 1;
+            return Position < dataBytes.Length - 1;
         }
 
         /// <summary>
@@ -166,16 +166,16 @@
                 return new CompactFontFormatData(Array.Empty<byte>());
             }
 
-            if (startLocation > dataBytes.Count - 1 || startLocation + length > dataBytes.Count)
+            if (startLocation > dataBytes.Length - 1 || startLocation + length > dataBytes.Length)
             {
-                throw new ArgumentException($"Attempted to create a snapshot of an invalid portion of the data. Length was {dataBytes.Count}, requested start: {startLocation} and requested length: {length}.");
+                throw new ArgumentException($"Attempted to create a snapshot of an invalid portion of the data. Length was {dataBytes.Length}, requested start: {startLocation} and requested length: {length}.");
             }
 
             var newData = new byte[length];
             var newI = 0;
             for (var i = startLocation; i < startLocation + length; i++)
             {
-                newData[newI] = dataBytes[i];
+                newData[newI] = dataBytes.Span[i];
                 newI++;
             }