--- a/src/UglyToad.PdfPig.Tests/Fonts/TrueType/TrueTypeDataBytesTests.cs
+++ b/src/UglyToad.PdfPig.Tests/Fonts/TrueType/TrueTypeDataBytesTests.cs
@@ -8,7 +8,7 @@
         [Fact]
         public void ReadUnsignedInt()
         {
-            var input = new ByteArrayInputBytes(new byte[]
+            var input = new MemoryInputBytes(new byte[]
             {
                 220,
                 43,