… src/UglyToad.PdfPig.Tests/Fonts/CompactFontFormat/Type2CharStringParserTests.cs
… src/UglyToad.PdfPig.Tests/Fonts/Parser/AdobeFontMetricsParserTests.cs
… src/UglyToad.PdfPig.Tests/Fonts/Parser/CMapParserTests.cs
… src/UglyToad.PdfPig.Tests/Fonts/TrueType/Parser/TrueTypeFontParserTests.cs
… src/UglyToad.PdfPig.Tests/Fonts/TrueType/Tables/Os2TableTests.cs
… src/UglyToad.PdfPig.Tests/Fonts/TrueType/TrueTypeChecksumCalculatorTests.cs
… src/UglyToad.PdfPig.Tests/Fonts/Type1/Type1FontParserTests.cs
… src/UglyToad.PdfPig.Tests/Graphics/Operations/GraphicsStateOperationTests.cs
… src/UglyToad.PdfPig.Tests/Graphics/TestOperationContext.cs
… src/UglyToad.PdfPig.Tests/IO/InputBytesTests.cs
… src/UglyToad.PdfPig.Tests/Images/PngFromPdfImageFactoryTests.cs
… src/UglyToad.PdfPig.Tests/Integration/AdvancedPdfDocumentAccessTests.cs
… src/UglyToad.PdfPig.Tests/Integration/EmbeddedFileAttachmentTests.cs
… src/UglyToad.PdfPig.Tests/Integration/PatternColorTests.cs
… src/UglyToad.PdfPig.Tests/Integration/SinglePageLibreOfficeImages.cs
… src/UglyToad.PdfPig.Tests/Integration/SwedishTouringCarChampionshipTests.cs
… src/UglyToad.PdfPig.Tests/Parser/Parts/BruteForceSearcherTests.cs
… src/UglyToad.PdfPig.Tests/Parser/Parts/FileHeaderParserTests.cs
… src/UglyToad.PdfPig.Tests/StringBytesTestConverter.cs
… src/UglyToad.PdfPig.Tests/TestPdfImage.cs
… src/UglyToad.PdfPig.Tests/Tokenization/StringTokenizerTests.cs
… src/UglyToad.PdfPig.Tests/Writer/Fonts/ToUnicodeCMapBuilderTests.cs
… src/UglyToad.PdfPig.Tests/Writer/PdfDocumentBuilderTests.cs
… src/UglyToad.PdfPig.Tests/Writer/PdfPageBuilderTests.cs
… src/UglyToad.PdfPig.Tokenization/Scanner/CoreTokenScanner.cs
… src/UglyToad.PdfPig.Tokens/InlineImageDataToken.cs
… src/UglyToad.PdfPig.Tokens/StreamToken.cs
… src/UglyToad.PdfPig/Annotations/AnnotationProvider.cs
… src/UglyToad.PdfPig/Annotations/QuadPointsQuadrilateral.cs
… src/UglyToad.PdfPig/Content/BasePageFactory.cs