This PR:

- Update the filters and their callers to utilize ReadOnly{Memory,Span}<byte>
- Polyfills Stream.Write(ReadOnlySpan<byte>)
- Updates various methods to accept ReadOnlySpan<byte>
- Eliminates various allocations
- Makes QuadPointsQuadrilateral a readonly struct
- Spanifies PngBuilder
- Spanifies InputBytes