@iamcarbon 2024-04-02
@BobLd Ready for review. This is the bulk of the spanification work and sets us up nicely for further optimization work.