@BobLd 2024-04-06
@iamcarbon I ran some benchmarks again following my previous comment. It seems the (alloc memory) performance decrease comes from your latest commit ([Remove unused StreamHelper methods](https://github.com/UglyToad/PdfPig/pull/815/commits/8314f411764d81ceb9f282214f63645b4b13b4c9))

memory2 (latest)
![image](https://github.com/UglyToad/PdfPig/assets/38405645/e412c7ab-2967-42d7-ac76-2d2c96df13a3)

memory2 (latest commit is `Eliminate allocation in BasePageFactory`)
![image](https://github.com/UglyToad/PdfPig/assets/38405645/a9e3622c-377e-44f8-8a19-54f2c46042c4)