@iamcarbon 2024-04-08
> @iamcarbon I ran some benchmarks again following my previous comment. It seems the (alloc memory) performance decrease comes from your latest commit ([Remove unused StreamHelper methods](https://github.com/UglyToad/PdfPig/pull/815/commits/8314f411764d81ceb9f282214f63645b4b13b4c9))
> 
> memory2 (latest) ![image](https://private-user-images.githubusercontent.com/38405645/320193929-e412c7ab-2967-42d7-ac76-2d2c96df13a3.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTI2MDgzMzMsIm5iZiI6MTcxMjYwODAzMywicGF0aCI6Ii8zODQwNTY0NS8zMjAxOTM5MjktZTQxMmM3YWItMjk2Ny00MmQ3LWFjNzYtMmQyYzk2ZGYxM2EzLnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA0MDglMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwNDA4VDIwMjcxM1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPWZjNzRiYzQ1YTgxMDJhNWExMjA3NTg0YTVkNzNmYmM2Y2U0OGZlNTFkYzM2YTRmNDQwNGE3YTg5MDM0YTk1MjkmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.x7utOZV6E_d4ll5MZ31y6ZKXABScst-shgybC2GJTfs)
> 
> memory2 (latest commit is `Eliminate allocation in BasePageFactory`) ![image](https://private-user-images.githubusercontent.com/38405645/320193930-a9e3622c-377e-44f8-8a19-54f2c46042c4.png?jwt=eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9.eyJpc3MiOiJnaXRodWIuY29tIiwiYXVkIjoicmF3LmdpdGh1YnVzZXJjb250ZW50LmNvbSIsImtleSI6ImtleTUiLCJleHAiOjE3MTI2MDgzMzMsIm5iZiI6MTcxMjYwODAzMywicGF0aCI6Ii8zODQwNTY0NS8zMjAxOTM5MzAtYTllMzYyMmMtMzc3ZS00NGY4LThhMTktNTRmMmM0NjA0MmM0LnBuZz9YLUFtei1BbGdvcml0aG09QVdTNC1ITUFDLVNIQTI1NiZYLUFtei1DcmVkZW50aWFsPUFLSUFWQ09EWUxTQTUzUFFLNFpBJTJGMjAyNDA0MDglMkZ1cy1lYXN0LTElMkZzMyUyRmF3czRfcmVxdWVzdCZYLUFtei1EYXRlPTIwMjQwNDA4VDIwMjcxM1omWC1BbXotRXhwaXJlcz0zMDAmWC1BbXotU2lnbmF0dXJlPTc3ZDcyMTZmOTY0ZDVjMGJjMzlhNzE3MjU5YjBlMWYyMGVmZDJlNjA1ZTk4YTlhNzU5MmI5NzVkNDllNGRmYjkmWC1BbXotU2lnbmVkSGVhZGVycz1ob3N0JmFjdG9yX2lkPTAma2V5X2lkPTAmcmVwb19pZD0wIn0.k2UiefGo2dXdZBe8GYEXyx6JLq9MfeP_dwTX3oIuBu8)

Will audit all our customer streams to see if there's a silent allocation somewhere. I changed our four WriteByte operations (which silently allocations on NET) to utilize a single Write operation, but I wouldn't expect that to allocate.