@BobLd 2024-04-08
This is the benchmark I ran

https://github.com/BobLd/Caly/blob/master/Caly.Benchmarks/OpenPdfBenchmark.cs

It does `GetImages()` which I think uses the 4 WriteBytes you replaced with a stackalloc

I agree with your remark though as I would not expect such a performance impact either...