@iamcarbon 2024-04-08
I'll profile this, and continue to work on recreating / fixing this. This has my curiosity. 

Can you also confirm what framework you were testing against?