@BobLd 2024-04-08
I was testing with dotnet8. I hope I didn't mess anything in the benchmark. Very curious too, let me know what you find