@iamcarbon 2024-04-12
@BobLd 

Oddly, I'm seeing a consistent ~215MB allocated before and after the StreamHelper changes, from both the memory benchmark, and from a heap capture.

BEFORE StreamHelper changes
```
| Method | Mean     | Error   | StdDev   | Ratio | Gen0       | Gen1       | Gen2      | Allocated | Alloc Ratio |
|------- |---------:|--------:|---------:|------:|-----------:|-----------:|----------:|----------:|------------:|
| PdfPig | 506.0 ms | 9.93 ms | 11.44 ms |  1.00 | 28000.0000 | 14000.0000 | 7000.0000 | 215.79 MB |        1.00 |
```

AFTER StreamHelper changes
```
| Method | Mean     | Error    | StdDev   | Ratio | Gen0       | Gen1       | Gen2      | Allocated | Alloc Ratio |
|------- |---------:|---------:|---------:|------:|-----------:|-----------:|----------:|----------:|------------:|
| PdfPig | 528.0 ms | 10.27 ms | 12.99 ms |  1.00 | 35000.0000 | 20000.0000 | 8000.0000 | 215.78 MB |        1.00 |
```

When inspecting the heap, my guess is that the additional usage is from the ArrayPool rooting some new buffers. These would improve steady state performance at the cost of higher base memory usage. The ArrayPool would also get purged if there's memory pressure. 

Curious if we should try to optimize this, or work on reducing allocations elsewhere, and making further use of the pool to amortize the cost of keeping these new buffers around.

--- 

Some captures of the current heap usage:

<img width="338" alt="Screenshot 2024-04-11 at 6 07 10 PM" src="https://github.com/UglyToad/PdfPig/assets/7537/8faa78fb-b0ab-440c-a1e3-52de19c91c2a">

<img width="1104" alt="Screenshot 2024-04-11 at 6 00 08 PM" src="https://github.com/UglyToad/PdfPig/assets/7537/d63ec3ec-144f-4225-b4d2-cb5ba3ab0ee7">

--

<img width="1130" alt="Screenshot 2024-04-11 at 6 00 17 PM" src="https://github.com/UglyToad/PdfPig/assets/7537/c6e4efe2-6b94-4fc3-8344-b962dc43949f">